komentari na zadatku se salju preko hub-a #135

See merge request si2024/syncatech!155

+++ src/back/backAPI/backAPI/Repositories/Implementation/Issues/IssueCommentRepository.cs
@@ -20,7 +20,7 @@ namespace backAPI.Repositories.Implementation.Issues
         {
             IssueComment ic = new IssueComment
             {
-                Id = (int)model.Id,
+                Id = model.Id,
                 UserId = Int32.Parse(model.UserId),
                 IssueId = Int32.Parse(model.IssueId),
                 Content = model.Body,
